Fix dropdown box size issue in Safari
The height of the dropdown box is shrinked to fit into the content size, ignoring the provided padding values in Safari. Adding height value to the dropdown box creates another issue of text getting cut in Chrome. As an alternative solution, height value is given to dropdown box and initial padding value is adjusted from 24px to 22px.